Create fib-i.s
Based on the figure on page 2 recursion is not the best way to handle this problem. There are
many calls that are repeated. As it turns out, the recursive Fibonacci implementation has a run
time of O (2n)! This problem should be linear O (n) (you can verify for yourself by writing down
the Fibonacci base cases and calculating the n-th number, you should notice right away you are
doing a single sweep of the numbers, hence linear runtime).